requirements.txt
# ...existing code...
flask~=3.1.1
boto3~=1.38.21
pillow~=11.2.1
openai~=1.81.0
requests~=2.32.3
# AI 코드에 필요한 패키지 추가 (예: torch, tensorflow 등)
# ...existing code...

pillow_heif~=0.22.0
botocore~=1.38.21